Moves apply_access_permissions call from base actor to generic_works_actor and modifies solr_helper to use edit_users as search index.
I had to move this method out of base actor due to the more
complex use of Delegate and Embargo.  The actions that govern
this functionality happen in the generic_works_actor, so the
permissions were not filtering down to the files attached.

The modification to the solr helper is to allow works to be
discoverable by both the depositor (delegate) and the owner.